spotless: rename core module after applying spotless
