+++ apps/mobile/build.gradle.kts
@@ -62,7 +62,7 @@ android {
 
 dependencies {
 
-  implementation(project(":features:core:catalog-source"))
+  implementation(project(":features:core:catalog"))
   implementation(project(":features:core:database"))
   implementation(project(":features:core:preferences-datastore"))
