+++ features/core/catalog/.gitignore
