+++ features/core/catalog/src/main/resources/en/catalog.json
