+++ features/core/catalog/src/test/resources/catalog-single.json
